feat(ui): extract sidebar component
